TCP KeyMatch command
The new `K` keys command can return arbitrary sized results, so a TCP mode was added to the server and client.

subtree has changed to red-black from AVL, due to being faster for inserts, and we do not necessarily need the perfect sorting of keys.

GC of namespaces will be a bit more consistent - a namespace will not be GC'd two passes in a row.